+++ src/components/hwpl/HwplItemSource/HwplItemSourceExpansionPanels.vue
@@ -1,6 +1,6 @@
 <script setup lang='ts'>
 import type { ItemSource } from '../../../types/HWPL/extra/ItemSource';
-import { getItemSource } from '../../../api/itemSource';
+import { getItemSource } from '../../../api/getItemSource';
 import asyncComputed from '../../../utils/asyncComputed';
 import HwplItemSourceGachaBoxes from './sources/HwplItemSourceGachaBoxes.vue';
 import HwplItemSourceEventAchievementRewards from './sources/HwplItemSourceEventAchievementRewards.vue';
