+++ src/views/Info/CharacterCards/CharacterCardList/helper/useSortCharacterCards.ts
@@ -3,7 +3,7 @@ import { computed, unref } from 'vue';
 import type { CharacterCard } from '@/types/HWPL/CharacterCard';
 import type { MaybeRef } from '@/types/vue/ref';
 import asyncComputed from '@/utils/asyncComputed';
-import { getItemSourceAggregateInformation } from '@/api/itemSource';
+import { getItemSourceAggregateInformation } from '@/api/getItemSource';
 import { useSort } from '@/composables/useSort';
 
 export type CharacterCardSortBy = 'Default' | 'Ability' | 'ReleaseTimestamp';
